backend/requirements.txt
fastapi
uvicorn
sqlalchemy
sendgrid
python-dotenv
pydantic
